feat: httpライブラリを分離してrssでもそちらを使うように変更
その他HTTPのリクエストにおいても同様に扱うべきであるため。
